app/src/main/java/ece651/sp22/xg73/battleship/NoCollisionRuleChecker.java
package ece651.sp22.xg73.battleship;

public class NoCollisionRuleChecker<T> extends PlacementRuleChecker<T> {

  public NoCollisionRuleChecker(PlacementRuleChecker<T> next) {
    super(next);
  }

  /**
   * Check the ship has no collision
   * 
   * @param The object of ship
   * @param the object of the board
   * @return A string indicate if the ship has collision with each others, if no collision happened, return null
   */

  @Override
  protected String checkMyRule(Ship<T> theShip, Board<T> theBoard) {
    for (Coordinate c : theShip.getCoordinates()) {
      if (theBoard.whatIsAtForSelf(c) != null) {
        return "That placement is invalid: the ship overlaps another ship.";
      }
    }
    return null;
  }
}
